refactor(state): extract signal management and reorganize tests
- Extract signal management to src/state/signals.rs (~200 lines)
  - Move SignalEvent struct, SIGNAL_QUEUE, enqueue_signal(), process_pending_signals()
  - Add comprehensive module documentation
  - Fix doctest imports to use correct paths

- Reorganize state tests into focused modules:
  - src/state/tests/fd_table_tests.rs - FD table operations (18 tests)
  - src/state/tests/options_tests.rs - shell options (17 tests)
  - src/state/tests/variable_tests.rs - variable scoping (8 tests)
  - src/state/tests/state_tests.rs - general state (10 tests)

- Update src/state/mod.rs:
  - Add signals module and re-exports
  - Remove all test code (~793 lines)
  - Add test module declaration
  - Reduced from ~1,540 to ~672 lines

All 669 tests pass. Phase 4 complete.
