+++ src/state/mod.rs
@@ -4,51 +4,22 @@ pub mod fd_table;
 // Shell options module
 pub mod options;
 
+// Signal management module
+pub mod signals;
+
 // Re-export types for backward compatibility
 pub use fd_table::FileDescriptorTable;
 pub use options::ShellOptions;
+pub use signals::{enqueue_signal, process_pending_signals};
 
 use super::parser::Ast;
-use lazy_static::lazy_static;
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet, VecDeque};
+use std::collections::{HashMap, HashSet};
 use std::env;
 use std::io::IsTerminal;
 use std::os::unix::io::RawFd;
 use std::rc::Rc;
 use std::sync::{Arc, Mutex};
-use std::time::Instant;
-
-lazy_static! {
-    /// Global queue for pending signal events
-    /// Signals are enqueued by the signal handler thread and dequeued by the main thread
-    pub static ref SIGNAL_QUEUE: Arc<Mutex<VecDeque<SignalEvent>>> =
-        Arc::new(Mutex::new(VecDeque::new()));
-}
-
-/// Maximum number of signals to queue before dropping old ones
-const MAX_SIGNAL_QUEUE_SIZE: usize = 100;
-
-/// Represents a signal event that needs to be processed
-#[derive(Debug, Clone)]
-pub struct SignalEvent {
-    /// Signal name (e.g., "INT", "TERM")
-    pub signal_name: String,
-    /// Signal number (e.g., 2, 15)
-    pub signal_number: i32,
-    /// When the signal was received
-    pub timestamp: Instant,
-}
-
-impl SignalEvent {
-    pub fn new(signal_name: String, signal_number: i32) -> Self {
-        Self {
-            signal_name,
-            signal_number,
-            timestamp: Instant::now(),
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct ColorScheme {
@@ -692,57 +663,6 @@ impl ShellState {
     }
 }
 
-/// Enqueue a signal event for later processing
-/// If the queue is full, the oldest event is dropped
-pub fn enqueue_signal(signal_name: &str, signal_number: i32) {
-    if let Ok(mut queue) = SIGNAL_QUEUE.lock() {
-        // If queue is full, remove oldest event
-        if queue.len() >= MAX_SIGNAL_QUEUE_SIZE {
-            queue.pop_front();
-            eprintln!("Warning: Signal queue overflow, dropping oldest signal");
-        }
-
-        queue.push_back(SignalEvent::new(signal_name.to_string(), signal_number));
-    }
-}
-
-/// Process all pending signals in the queue
-/// This should be called at safe points during command execution
-pub fn process_pending_signals(shell_state: &mut ShellState) {
-    // Try to lock the queue with a timeout to avoid blocking
-    if let Ok(mut queue) = SIGNAL_QUEUE.lock() {
-        // Process all pending signals
-        while let Some(signal_event) = queue.pop_front() {
-            // Check if a trap is set for this signal
-            if let Some(trap_cmd) = shell_state.get_trap(&signal_event.signal_name)
-                && !trap_cmd.is_empty()
-            {
-                // Display signal information for debugging/tracking
-                if shell_state.colors_enabled {
-                    eprintln!(
-                        "{}Signal {} (signal {}) received at {:?}\x1b[0m",
-                        shell_state.color_scheme.builtin,
-                        signal_event.signal_name,
-                        signal_event.signal_number,
-                        signal_event.timestamp
-                    );
-                } else {
-                    eprintln!(
-                        "Signal {} (signal {}) received at {:?}",
-                        signal_event.signal_name,
-                        signal_event.signal_number,
-                        signal_event.timestamp
-                    );
-                }
-
-                // Execute the trap handler
-                // Note: This preserves the exit code as per POSIX requirements
-                crate::executor::execute_trap_handler(&trap_cmd, shell_state);
-            }
-        }
-    }
-}
-
 impl Default for ShellState {
     fn default() -> Self {
         Self::new()
@@ -750,796 +670,4 @@ impl Default for ShellState {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::{fs::File, os::fd::AsRawFd, sync::Mutex};
-
-    // Mutex to serialize tests that create temporary files
-    static FILE_LOCK: Mutex<()> = Mutex::new(());
-
-    #[test]
-    fn test_shell_state_basic() {
-        let mut state = ShellState::new();
-        state.set_var("TEST_VAR", "test_value".to_string());
-        assert_eq!(state.get_var("TEST_VAR"), Some("test_value".to_string()));
-    }
-
-    #[test]
-    fn test_special_variables() {
-        let mut state = ShellState::new();
-        state.set_last_exit_code(42);
-        state.set_script_name("test_script");
-
-        assert_eq!(state.get_var("?"), Some("42".to_string()));
-        assert_eq!(state.get_var("$"), Some(state.shell_pid.to_string()));
-        assert_eq!(state.get_var("0"), Some("test_script".to_string()));
-    }
-
-    #[test]
-    fn test_export_variable() {
-        let mut state = ShellState::new();
-        state.set_var("EXPORT_VAR", "export_value".to_string());
-        state.export_var("EXPORT_VAR");
-
-        let child_env = state.get_env_for_child();
-        assert_eq!(
-            child_env.get("EXPORT_VAR"),
-            Some(&"export_value".to_string())
-        );
-    }
-
-    #[test]
-    fn test_unset_variable() {
-        let mut state = ShellState::new();
-        state.set_var("UNSET_VAR", "value".to_string());
-        state.export_var("UNSET_VAR");
-
-        assert!(state.variables.contains_key("UNSET_VAR"));
-        assert!(state.exported.contains("UNSET_VAR"));
-
-        state.unset_var("UNSET_VAR");
-
-        assert!(!state.variables.contains_key("UNSET_VAR"));
-        assert!(!state.exported.contains("UNSET_VAR"));
-    }
-
-    #[test]
-    fn test_get_user_hostname() {
-        let state = ShellState::new();
-        let user_hostname = state.get_user_hostname();
-        // Should contain @ since it's user@hostname format
-        assert!(user_hostname.contains('@'));
-    }
-
-    #[test]
-    fn test_get_prompt() {
-        let state = ShellState::new();
-        let prompt = state.get_prompt();
-        // Should end with $ and contain @
-        assert!(prompt.ends_with(" $ "));
-        assert!(prompt.contains('@'));
-    }
-
-    #[test]
-    fn test_positional_parameters() {
-        let mut state = ShellState::new();
-        state.set_positional_params(vec![
-            "arg1".to_string(),
-            "arg2".to_string(),
-            "arg3".to_string(),
-        ]);
-
-        assert_eq!(state.get_var("1"), Some("arg1".to_string()));
-        assert_eq!(state.get_var("2"), Some("arg2".to_string()));
-        assert_eq!(state.get_var("3"), Some("arg3".to_string()));
-        assert_eq!(state.get_var("4"), None);
-        assert_eq!(state.get_var("#"), Some("3".to_string()));
-        assert_eq!(state.get_var("*"), Some("arg1 arg2 arg3".to_string()));
-        assert_eq!(state.get_var("@"), Some("arg1 arg2 arg3".to_string()));
-    }
-
-    #[test]
-    fn test_positional_parameters_empty() {
-        let mut state = ShellState::new();
-        state.set_positional_params(vec![]);
-
-        assert_eq!(state.get_var("1"), None);
-        assert_eq!(state.get_var("#"), Some("0".to_string()));
-        assert_eq!(state.get_var("*"), Some("".to_string()));
-        assert_eq!(state.get_var("@"), Some("".to_string()));
-    }
-
-    #[test]
-    fn test_shift_positional_params() {
-        let mut state = ShellState::new();
-        state.set_positional_params(vec![
-            "arg1".to_string(),
-            "arg2".to_string(),
-            "arg3".to_string(),
-        ]);
-
-        assert_eq!(state.get_var("1"), Some("arg1".to_string()));
-        assert_eq!(state.get_var("2"), Some("arg2".to_string()));
-        assert_eq!(state.get_var("3"), Some("arg3".to_string()));
-
-        state.shift_positional_params(1);
-
-        assert_eq!(state.get_var("1"), Some("arg2".to_string()));
-        assert_eq!(state.get_var("2"), Some("arg3".to_string()));
-        assert_eq!(state.get_var("3"), None);
-        assert_eq!(state.get_var("#"), Some("2".to_string()));
-
-        state.shift_positional_params(2);
-
-        assert_eq!(state.get_var("1"), None);
-        assert_eq!(state.get_var("#"), Some("0".to_string()));
-    }
-
-    #[test]
-    fn test_push_positional_param() {
-        let mut state = ShellState::new();
-        state.set_positional_params(vec!["arg1".to_string()]);
-
-        assert_eq!(state.get_var("1"), Some("arg1".to_string()));
-        assert_eq!(state.get_var("#"), Some("1".to_string()));
-
-        state.push_positional_param("arg2".to_string());
-
-        assert_eq!(state.get_var("1"), Some("arg1".to_string()));
-        assert_eq!(state.get_var("2"), Some("arg2".to_string()));
-        assert_eq!(state.get_var("#"), Some("2".to_string()));
-    }
-
-    #[test]
-    fn test_local_variable_scoping() {
-        let mut state = ShellState::new();
-
-        // Set a global variable
-        state.set_var("global_var", "global_value".to_string());
-        assert_eq!(
-            state.get_var("global_var"),
-            Some("global_value".to_string())
-        );
-
-        // Push local scope
-        state.push_local_scope();
-
-        // Set a local variable with the same name
-        state.set_local_var("global_var", "local_value".to_string());
-        assert_eq!(state.get_var("global_var"), Some("local_value".to_string()));
-
-        // Set another local variable
-        state.set_local_var("local_var", "local_only".to_string());
-        assert_eq!(state.get_var("local_var"), Some("local_only".to_string()));
-
-        // Pop local scope
-        state.pop_local_scope();
-
-        // Should be back to global variable
-        assert_eq!(
-            state.get_var("global_var"),
-            Some("global_value".to_string())
-        );
-        assert_eq!(state.get_var("local_var"), None);
-    }
-
-    #[test]
-    fn test_nested_local_scopes() {
-        let mut state = ShellState::new();
-
-        // Set global variable
-        state.set_var("test_var", "global".to_string());
-
-        // Push first local scope
-        state.push_local_scope();
-        state.set_local_var("test_var", "level1".to_string());
-        assert_eq!(state.get_var("test_var"), Some("level1".to_string()));
-
-        // Push second local scope
-        state.push_local_scope();
-        state.set_local_var("test_var", "level2".to_string());
-        assert_eq!(state.get_var("test_var"), Some("level2".to_string()));
-
-        // Pop second scope
-        state.pop_local_scope();
-        assert_eq!(state.get_var("test_var"), Some("level1".to_string()));
-
-        // Pop first scope
-        state.pop_local_scope();
-        assert_eq!(state.get_var("test_var"), Some("global".to_string()));
-    }
-
-    #[test]
-    fn test_variable_set_in_local_scope() {
-        let mut state = ShellState::new();
-
-        // No local scope initially
-        state.set_var("test_var", "global".to_string());
-        assert_eq!(state.get_var("test_var"), Some("global".to_string()));
-
-        // Push local scope and set local variable
-        state.push_local_scope();
-        state.set_local_var("test_var", "local".to_string());
-        assert_eq!(state.get_var("test_var"), Some("local".to_string()));
-
-        // Pop scope
-        state.pop_local_scope();
-        assert_eq!(state.get_var("test_var"), Some("global".to_string()));
-    }
-
-    #[test]
-    fn test_condensed_cwd_environment_variable() {
-        // Test default behavior (should be true for backward compatibility)
-        let state = ShellState::new();
-        assert!(state.condensed_cwd);
-
-        // Test explicit true
-        unsafe {
-            env::set_var("RUSH_CONDENSED", "true");
-        }
-        let state = ShellState::new();
-        assert!(state.condensed_cwd);
-
-        // Test explicit false
-        unsafe {
-            env::set_var("RUSH_CONDENSED", "false");
-        }
-        let state = ShellState::new();
-        assert!(!state.condensed_cwd);
-
-        // Clean up
-        unsafe {
-            env::remove_var("RUSH_CONDENSED");
-        }
-    }
-
-    #[test]
-    fn test_get_full_cwd() {
-        let state = ShellState::new();
-        let full_cwd = state.get_full_cwd();
-        assert!(!full_cwd.is_empty());
-        // Should contain path separators (either / or \ depending on platform)
-        assert!(full_cwd.contains('/') || full_cwd.contains('\\'));
-    }
-
-    #[test]
-    fn test_prompt_with_condensed_setting() {
-        let mut state = ShellState::new();
-
-        // Test with condensed enabled (default)
-        assert!(state.condensed_cwd);
-        let prompt_condensed = state.get_prompt();
-        assert!(prompt_condensed.contains('@'));
-
-        // Test with condensed disabled
-        state.condensed_cwd = false;
-        let prompt_full = state.get_prompt();
-        assert!(prompt_full.contains('@'));
-
-        // Both should end with "$ " (or "# " for root)
-        assert!(prompt_condensed.ends_with("$ ") || prompt_condensed.ends_with("# "));
-        assert!(prompt_full.ends_with("$ ") || prompt_full.ends_with("# "));
-    }
-
-    // File Descriptor Table Tests
-
-    #[test]
-    fn test_fd_table_creation() {
-        let fd_table = FileDescriptorTable::new();
-        assert!(!fd_table.is_open(0));
-        assert!(!fd_table.is_open(1));
-        assert!(!fd_table.is_open(2));
-    }
-
-    #[test]
-    fn test_fd_table_open_file() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_open.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file for reading
-        let result = fd_table.open_fd(3, temp_file, true, false, false, false, false);
-        assert!(result.is_ok());
-        assert!(fd_table.is_open(3));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_open_file_for_writing() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file path
-        let temp_file = "/tmp/rush_test_fd_write.txt";
-
-        // Open file for writing
-        let result = fd_table.open_fd(4, temp_file, false, true, false, true, false);
-        assert!(result.is_ok());
-        assert!(fd_table.is_open(4));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_invalid_fd_number() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Test invalid fd numbers
-        let result = fd_table.open_fd(-1, "/tmp/test.txt", true, false, false, false, false);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Invalid file descriptor"));
-
-        let result = fd_table.open_fd(1025, "/tmp/test.txt", true, false, false, false, false);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Invalid file descriptor"));
-    }
-
-    #[test]
-    fn test_fd_table_duplicate_fd() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_dup.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file on fd 3
-        fd_table
-            .open_fd(3, temp_file, true, false, false, false, false)
-            .unwrap();
-        assert!(fd_table.is_open(3));
-
-        // Duplicate fd 3 to fd 4
-        let result = fd_table.duplicate_fd(3, 4);
-        assert!(result.is_ok());
-        assert!(fd_table.is_open(4));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_duplicate_to_self() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_dup_self.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file on fd 3
-        fd_table
-            .open_fd(3, temp_file, true, false, false, false, false)
-            .unwrap();
-
-        // Duplicate fd 3 to itself (should be no-op)
-        let result = fd_table.duplicate_fd(3, 3);
-        assert!(result.is_ok());
-        assert!(fd_table.is_open(3));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_duplicate_closed_fd() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Try to duplicate a closed fd
-        let result = fd_table.duplicate_fd(3, 4);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("not open"));
-    }
-
-    #[test]
-    fn test_fd_table_close_fd() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_close.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file on fd 3
-        fd_table
-            .open_fd(3, temp_file, true, false, false, false, false)
-            .unwrap();
-        assert!(fd_table.is_open(3));
-
-        // Close fd 3
-        let result = fd_table.close_fd(3);
-        assert!(result.is_ok());
-        assert!(fd_table.is_closed(3));
-        assert!(!fd_table.is_open(3));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_save_and_restore() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Save stdin (fd 0)
-        let result = fd_table.save_fd(0);
-        assert!(result.is_ok());
-
-        // Restore stdin
-        let result = fd_table.restore_fd(0);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_fd_table_save_all_and_restore_all() {
-        let _lock = FILE_LOCK.lock().unwrap();
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create unique temporary files
-        use std::time::{SystemTime, UNIX_EPOCH};
-        let timestamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap()
-            .as_nanos();
-        let temp_file1 = format!("/tmp/rush_test_fd_save1_{}.txt", timestamp);
-        let temp_file2 = format!("/tmp/rush_test_fd_save2_{}.txt", timestamp);
-
-        std::fs::write(&temp_file1, "test content 1").unwrap();
-        std::fs::write(&temp_file2, "test content 2").unwrap();
-
-        // Open files on fd 50 and 51
-        // Manually dup2 to ensure these FDs are valid for save_fd()
-        // Using higher FDs to avoid conflict with parallel tests using 0-9
-        let f1 = File::open(&temp_file1).unwrap();
-        let f2 = File::open(&temp_file2).unwrap();
-        unsafe {
-            libc::dup2(f1.as_raw_fd(), 50);
-            libc::dup2(f2.as_raw_fd(), 51);
-        }
-
-        fd_table
-            .open_fd(50, &temp_file1, true, false, false, false, false)
-            .unwrap();
-        fd_table
-            .open_fd(51, &temp_file2, true, false, false, false, false)
-            .unwrap();
-
-        // Save all fds
-        let result = fd_table.save_all_fds();
-        assert!(result.is_ok());
-
-        // Restore all fds
-        let result = fd_table.restore_all_fds();
-        assert!(result.is_ok());
-
-        // Clean up
-        unsafe {
-            libc::close(50);
-            libc::close(51);
-        }
-        let _ = std::fs::remove_file(&temp_file1);
-        let _ = std::fs::remove_file(&temp_file2);
-    }
-
-    #[test]
-    fn test_fd_table_clear() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_clear.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file on fd 50 (was 3)
-        // Manual setup not strictly needed for clear() test as it checks map?
-        // But clear() might close FDs?
-        // FileDescriptorTable::clear() just clears map. File drops.
-
-        fd_table
-            .open_fd(50, temp_file, true, false, false, false, false)
-            .unwrap();
-        assert!(fd_table.is_open(50));
-
-        // Clear all fds
-        fd_table.clear();
-        assert!(!fd_table.is_open(3));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_get_stdio() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_fd_stdio.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file on fd 3
-        fd_table
-            .open_fd(3, temp_file, true, false, false, false, false)
-            .unwrap();
-
-        // Get Stdio for fd 3
-        let stdio = fd_table.get_stdio(3);
-        assert!(stdio.is_some());
-
-        // Get Stdio for non-existent fd
-        let stdio = fd_table.get_stdio(5);
-        assert!(stdio.is_none());
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    #[test]
-    fn test_fd_table_multiple_operations() {
-        let mut fd_table = FileDescriptorTable::new();
-
-        // Create temporary files
-        let temp_file1 = "/tmp/rush_test_fd_multi1.txt";
-        let temp_file2 = "/tmp/rush_test_fd_multi2.txt";
-        std::fs::write(temp_file1, "test content 1").unwrap();
-        std::fs::write(temp_file2, "test content 2").unwrap();
-
-        // Open file on fd 3
-        fd_table
-            .open_fd(3, temp_file1, true, false, false, false, false)
-            .unwrap();
-        assert!(fd_table.is_open(3));
-
-        // Duplicate fd 3 to fd 4
-        fd_table.duplicate_fd(3, 4).unwrap();
-        assert!(fd_table.is_open(4));
-
-        // Open another file on fd 5
-        fd_table
-            .open_fd(5, temp_file2, true, false, false, false, false)
-            .unwrap();
-        assert!(fd_table.is_open(5));
-
-        // Close fd 4
-        fd_table.close_fd(4).unwrap();
-        assert!(fd_table.is_closed(4));
-        assert!(!fd_table.is_open(4));
-
-        // fd 3 and 5 should still be open
-        assert!(fd_table.is_open(3));
-        assert!(fd_table.is_open(5));
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file1);
-        let _ = std::fs::remove_file(temp_file2);
-    }
-
-    #[test]
-    fn test_shell_state_has_fd_table() {
-        let state = ShellState::new();
-        let fd_table = state.fd_table.borrow();
-        assert!(!fd_table.is_open(3));
-    }
-
-    #[test]
-    fn test_shell_state_fd_table_operations() {
-        let state = ShellState::new();
-
-        // Create a temporary file
-        let temp_file = "/tmp/rush_test_state_fd.txt";
-        std::fs::write(temp_file, "test content").unwrap();
-
-        // Open file through shell state's fd table
-        {
-            let mut fd_table = state.fd_table.borrow_mut();
-            fd_table
-                .open_fd(3, temp_file, true, false, false, false, false)
-                .unwrap();
-        }
-
-        // Verify it's open
-        {
-            let fd_table = state.fd_table.borrow();
-            assert!(fd_table.is_open(3));
-        }
-
-        // Clean up
-        let _ = std::fs::remove_file(temp_file);
-    }
-
-    // ShellOptions Tests
-
-    #[test]
-    fn test_shell_options_default() {
-        let options = ShellOptions::default();
-        assert!(!options.errexit);
-        assert!(!options.nounset);
-        assert!(!options.xtrace);
-        assert!(!options.verbose);
-        assert!(!options.noexec);
-        assert!(!options.noglob);
-        assert!(!options.noclobber);
-        assert!(!options.allexport);
-        assert!(!options.notify);
-        assert!(!options.ignoreeof);
-        assert!(!options.monitor);
-    }
-
-    #[test]
-    fn test_shell_options_get_by_short_name() {
-        let mut options = ShellOptions::default();
-        options.errexit = true;
-        options.nounset = true;
-
-        assert_eq!(options.get_by_short_name('e'), Some(true));
-        assert_eq!(options.get_by_short_name('u'), Some(true));
-        assert_eq!(options.get_by_short_name('x'), Some(false));
-        assert_eq!(options.get_by_short_name('Z'), None);
-    }
-
-    #[test]
-    fn test_shell_options_set_by_short_name() {
-        let mut options = ShellOptions::default();
-
-        assert!(options.set_by_short_name('e', true).is_ok());
-        assert!(options.errexit);
-
-        assert!(options.set_by_short_name('u', true).is_ok());
-        assert!(options.nounset);
-
-        assert!(options.set_by_short_name('x', true).is_ok());
-        assert!(options.xtrace);
-
-        assert!(options.set_by_short_name('e', false).is_ok());
-        assert!(!options.errexit);
-
-        // Invalid option
-        assert!(options.set_by_short_name('Z', true).is_err());
-    }
-
-    #[test]
-    fn test_shell_options_get_by_long_name() {
-        let mut options = ShellOptions::default();
-        options.errexit = true;
-        options.nounset = true;
-
-        assert_eq!(options.get_by_long_name("errexit"), Some(true));
-        assert_eq!(options.get_by_long_name("nounset"), Some(true));
-        assert_eq!(options.get_by_long_name("xtrace"), Some(false));
-        assert_eq!(options.get_by_long_name("invalid"), None);
-    }
-
-    #[test]
-    fn test_shell_options_set_by_long_name() {
-        let mut options = ShellOptions::default();
-
-        assert!(options.set_by_long_name("errexit", true).is_ok());
-        assert!(options.errexit);
-
-        assert!(options.set_by_long_name("nounset", true).is_ok());
-        assert!(options.nounset);
-
-        assert!(options.set_by_long_name("xtrace", true).is_ok());
-        assert!(options.xtrace);
-
-        assert!(options.set_by_long_name("errexit", false).is_ok());
-        assert!(!options.errexit);
-
-        // Invalid option
-        assert!(options.set_by_long_name("invalid", true).is_err());
-    }
-
-    #[test]
-    fn test_shell_options_all_short_options() {
-        let mut options = ShellOptions::default();
-
-        // Test all valid short options
-        let short_opts = vec!['e', 'u', 'x', 'v', 'n', 'f', 'C', 'a', 'b', 'm'];
-        for opt in short_opts {
-            assert!(options.set_by_short_name(opt, true).is_ok());
-            assert_eq!(options.get_by_short_name(opt), Some(true));
-            assert!(options.set_by_short_name(opt, false).is_ok());
-            assert_eq!(options.get_by_short_name(opt), Some(false));
-        }
-    }
-
-    #[test]
-    fn test_shell_options_all_long_options() {
-        let mut options = ShellOptions::default();
-
-        // Test all valid long options
-        let long_opts = vec![
-            "errexit",
-            "nounset",
-            "xtrace",
-            "verbose",
-            "noexec",
-            "noglob",
-            "noclobber",
-            "allexport",
-            "notify",
-            "ignoreeof",
-            "monitor",
-        ];
-        for opt in long_opts {
-            assert!(options.set_by_long_name(opt, true).is_ok());
-            assert_eq!(options.get_by_long_name(opt), Some(true));
-            assert!(options.set_by_long_name(opt, false).is_ok());
-            assert_eq!(options.get_by_long_name(opt), Some(false));
-        }
-    }
-
-    #[test]
-    fn test_shell_options_get_all_options() {
-        let mut options = ShellOptions::default();
-        options.errexit = true;
-        options.xtrace = true;
-
-        let all_options = options.get_all_options();
-
-        // Should have 11 options
-        assert_eq!(all_options.len(), 11);
-
-        // Find errexit and verify it's on
-        let errexit_opt = all_options.iter().find(|(name, _, _)| *name == "errexit");
-        assert!(errexit_opt.is_some());
-        assert_eq!(errexit_opt.unwrap().2, true);
-
-        // Find xtrace and verify it's on
-        let xtrace_opt = all_options.iter().find(|(name, _, _)| *name == "xtrace");
-        assert!(xtrace_opt.is_some());
-        assert_eq!(xtrace_opt.unwrap().2, true);
-
-        // Find nounset and verify it's off
-        let nounset_opt = all_options.iter().find(|(name, _, _)| *name == "nounset");
-        assert!(nounset_opt.is_some());
-        assert_eq!(nounset_opt.unwrap().2, false);
-    }
-
-    #[test]
-    fn test_shell_state_has_options() {
-        let state = ShellState::new();
-        assert!(!state.options.errexit);
-        assert!(!state.options.nounset);
-        assert!(!state.options.xtrace);
-    }
-
-    #[test]
-    fn test_shell_state_options_modification() {
-        let mut state = ShellState::new();
-
-        state.options.errexit = true;
-        assert!(state.options.errexit);
-
-        state.options.set_by_short_name('u', true).unwrap();
-        assert!(state.options.nounset);
-
-        state.options.set_by_long_name("xtrace", true).unwrap();
-        assert!(state.options.xtrace);
-    }
-
-    #[test]
-    fn test_shell_options_error_messages() {
-        let mut options = ShellOptions::default();
-
-        let result = options.set_by_short_name('Z', true);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Invalid option: -Z"));
-
-        let result = options.set_by_long_name("invalid_option", true);
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("Invalid option: invalid_option")
-        );
-    }
-
-    #[test]
-    fn test_shell_options_case_sensitivity() {
-        let mut options = ShellOptions::default();
-
-        // 'C' is valid (noclobber), 'c' is not
-        assert!(options.set_by_short_name('C', true).is_ok());
-        assert!(options.noclobber);
-        assert!(options.set_by_short_name('c', true).is_err());
-    }
-}
+mod tests;
