+++ src/state/tests/mod.rs
@@ -1,0 +1,12 @@
+//! Tests for the state module
+//!
+//! This module organizes tests into focused submodules:
+//! - `fd_table_tests`: File descriptor table operations
+//! - `options_tests`: Shell options management
+//! - `variable_tests`: Variable scoping and management
+//! - `state_tests`: General state management
+
+mod fd_table_tests;
+mod options_tests;
+mod state_tests;
+mod variable_tests;
